requirements.txt
langchain
langchain-core
langchain-community
openai
tiktoken
faiss-cpu
transformers
python-dotenv
streamlit